+++ app/gui/src/components/AppContainer/reactTabs.ts
@@ -3,12 +3,10 @@ import { AssetVersions as ReactAssetVersions } from '#/layouts/AssetPanel/compon
 import { ProjectExecutionsCalendar as ReactProjectExecutionsCalendar } from '#/layouts/AssetPanel/components/ProjectExecutionsCalendar'
 import { ProjectSessions as ReactProjectSessions } from '#/layouts/AssetPanel/components/ProjectSessions'
 import { Drive as ReactDrive } from '#/layouts/Drive'
-import ReactEditor from '#/layouts/Editor'
 import ReactSettings from '#/layouts/Settings'
 import { suspendedReactComponent } from '@/util/react'
 
 export const Drive = suspendedReactComponent(ReactDrive)
-export const Editor = suspendedReactComponent(ReactEditor)
 export const Settings = suspendedReactComponent(ReactSettings)
 export const AssetProperties = suspendedReactComponent(ReactAssetProperties)
 export const AssetVersions = suspendedReactComponent(ReactAssetVersions)