+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - [Many CLI arguments removed][14069]
 - [JSON and SQL visualizations' content may be now selected and copied][14262]
 - [SQL visualization displays interpolated parameters properly][14262]
+- [Multiple opened projects' tabs are now allowed][14215]
 
 [13685]: https://github.com/enso-org/enso/pull/13685
 [13658]: https://github.com/enso-org/enso/pull/13658
@@ -29,6 +30,7 @@
 [14059]: https://github.com/enso-org/enso/pull/14059
 [14069]: https://github.com/enso-org/enso/pull/14069
 [14262]: https://github.com/enso-org/enso/pull/14262
+[14215]: https://github.com/enso-org/enso/pull/14215
 
 #### Enso Standard Library
 
@@ -46,6 +48,7 @@
 - [Snowflake stage support for reading and writing files.][14210]
 - [Snowflake file format and copy into support.][14221]
 - [Snowflake bulk load API.][14230]
+- [Add error_on_missing_columns to more methods.][14236]
 
 [13769]: https://github.com/enso-org/enso/pull/13769
 [14026]: https://github.com/enso-org/enso/pull/14026
@@ -60,6 +63,7 @@
 [14210]: https://github.com/enso-org/enso/pull/14210
 [14221]: https://github.com/enso-org/enso/pull/14221
 [14230]: https://github.com/enso-org/enso/pull/14230
+[14236]: https://github.com/enso-org/enso/pull/14236
 
 #### Enso Language & Runtime